Retire DocumentFolderService back-compat shim (#1680)
PR #1685 split the monolithic DocumentFolderService along its natural
seam — DocumentService for document-level operations (quota, validation,
creation, lookup, permissions) and CorpusObjsService for corpus-scoped
operations (folder CRUD, document-to-corpus membership, soft-delete
lifecycle, the new IDOR-safe convenience methods). The original class
survived as a multiple-inheritance shim in folder_service.py so the ~18
existing importers could migrate later. This change finishes that
migration and deletes the shim per the no-dead-code rule in CLAUDE.md.

Production code (9 files): every DocumentFolderService.X() call rewritten
to CorpusObjsService.X(). Routing follows the partition pinned by
test_service_split_method_partition — none of these call sites hit a
DocumentService-routed method, so the rewrites are uniform across the
config/graphql/* mutation + query + type modules, the zip-import task,
the move-document agent tool, the MCP server, and the v2 import helper.

Tests (3 sibling files): same rewrite in test_permanent_deletion.py,
test_zip_import_integration.py, and permissioning/test_permission_optimizer.py.

Big test merge: test_document_folder_service.py (4423 lines) and the
existing test_corpus_objs_service.py (523 lines) collapsed into a single
test_corpus_objs_service.py (4925 lines). The two test classes that
exercise DocumentService methods (TestEdgeCases_UploadQuota,
TestValidateFileType) now call DocumentService directly and inherit
from TransactionTestCase rather than the folder-test base. The
folder-test base itself renamed from DocumentFolderServiceTestBase to
the underscore-prefixed _CorpusObjsServiceFolderTestBase so it no
longer clashes with the existing public CorpusObjsServiceTestBase
fixture for the convenience-method tests.

Shim removed: folder_service.py deleted; the TestServiceSplit_BackwardsCompatAlias
class in test_document_service.py and the corresponding docstring
bullet retired. The disjoint-partition test
(TestServiceSplit_MethodPartition) survives — it still pins the
invariant that DocumentService.__dict__ and CorpusObjsService.__dict__
share no method names.

Stale docstring sweep: comment / docstring mentions in corpuses/models.py,
corpuses/corpus_objs_service.py, mcp/resources.py, the three
DocumentFolderService-named test_* modules that just had passing
comments, the CLAUDE.md follow-up bullet at line 245, and the active
architecture / permissioning docs (bulk-import.md,
data_model/documents_and_annotations.md, REMEDIATION_GUIDE.md,
documentation_alignment_review.md) reworded to name the new services.
The three historical standalone docs
(docs/architecture/document_folder_service{,_plan,_migration_inventory}.md)
and their mkdocs nav entry deleted — they document a class that no
longer exists. docs/releases/v3.0.0.b3.md is intentionally untouched
(release notes are an immutable snapshot of the release that shipped);
docs/typing/mypy_baseline.txt carries stale entries that point at the
renamed test file but will be regenerated by the next mypy run.

Closes #1680
