+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- **Retire `DocumentFolderService` back-compat shim — migrate all importers to `DocumentService` / `CorpusObjsService`** (issue #1680). PR #1685 split the monolithic `DocumentFolderService` along its natural seam and kept the original class as a `class DocumentFolderService(CorpusObjsService, DocumentService)` shim in `opencontractserver/corpuses/folder_service.py` so the ~18 existing importers could migrate later. This change finishes that migration:
+  - **Production call sites (9 files):** `config/graphql/{corpus_folder_mutations,corpus_mutations,corpus_queries,document_mutations,document_types}.py`, `opencontractserver/tasks/import_tasks.py`, `opencontractserver/llms/tools/core_tools/documents.py`, `opencontractserver/mcp/server.py`, `opencontractserver/utils/import_v2.py` — every `DocumentFolderService.X(...)` rewritten to `CorpusObjsService.X(...)` based on the partition documented in `test_service_split_method_partition` (`opencontractserver/tests/test_corpus_objs_service.py:46`). None of these call sites hit a `DocumentService`-routed method.
+  - **Test call sites (3 files):** `opencontractserver/tests/test_permanent_deletion.py`, `opencontractserver/tests/test_zip_import_integration.py`, `opencontractserver/tests/permissioning/test_permission_optimizer.py` — same surgical rewrite to `CorpusObjsService`.
+  - **Test suite merge:** `opencontractserver/tests/test_document_folder_service.py` (4423 lines) and the existing `opencontractserver/tests/test_corpus_objs_service.py` (523 lines) collapsed into a single `opencontractserver/tests/test_corpus_objs_service.py` (4925 lines). The two test classes that exercise `DocumentService` methods (`TestEdgeCases_UploadQuota`, `TestValidateFileType`) now call `DocumentService.check_user_upload_quota` / `DocumentService.validate_file_type` directly and inherit from `TransactionTestCase` rather than the folder-test base. The folder-test base itself was renamed from `DocumentFolderServiceTestBase` to the underscore-prefixed `_CorpusObjsServiceFolderTestBase` so it no longer clashes with the existing public `CorpusObjsServiceTestBase` fixture for the convenience-method tests.
+  - **Shim removed:** `opencontractserver/corpuses/folder_service.py` deleted along with the `TestServiceSplit_BackwardsCompatAlias` class in `opencontractserver/tests/test_document_service.py` and the surrounding docstring bullet that explained the alias. The disjoint-partition test (`TestServiceSplit_MethodPartition`) survives — it still pins the invariant that `DocumentService.__dict__` and `CorpusObjsService.__dict__` share no method names, which is exactly what made the two-parent inheritance unambiguous in the first place and remains the load-bearing assertion behind the split.
+  - **Stale docstring sweep:** comment / docstring mentions of "DocumentFolderService" / "folder_service" in `opencontractserver/corpuses/{models.py,corpus_objs_service.py}`, `opencontractserver/mcp/resources.py`, `opencontractserver/tests/{test_document_stats.py,test_import_utils.py,test_move_document_tool.py}`, the `CLAUDE.md` follow-up bullet at line 245, and the active architecture / permissioning docs (`docs/architecture/{bulk-import.md,data_model/documents_and_annotations.md}`, `docs/permissioning/REMEDIATION_GUIDE.md`, `docs/plans/documentation_alignment_review.md`) reworded to name the new services. The three historical standalone docs (`docs/architecture/document_folder_service{,_plan,_migration_inventory}.md`) and the mkdocs nav entry pointing at the first of those deleted along with the shim — they document a class that no longer exists. `docs/releases/v3.0.0.b3.md` is intentionally untouched (release notes are an immutable snapshot of the release that shipped); `docs/typing/mypy_baseline.txt` carries stale entries that point at the renamed test file but will be regenerated by the next mypy run.
+
 - **Phase C permission centralization — migrate optimizer call sites to `obj.user_can(...)`** (issue #1657; `opencontractserver/annotations/query_optimizer.py`, `opencontractserver/extracts/query_optimizer.py`). The 24 remaining `user_has_permission_for_obj` call sites across `AnnotationQueryOptimizer` (analysis/extract/datacell/corpus-annotation/document-relationship paths) and `MetadataQueryOptimizer` (`_compute_effective_permissions`, `get_corpus_metadata_columns`, `get_documents_metadata_batch`, `check_metadata_mutation_permission`) now route through `obj.user_can(...)` — completing the substitution PR #1663 started in `AnnotationQueryOptimizer._compute_effective_permissions`. Net effect:
   - **Cache plumbing.** Three of the migrated sites (`AnnotationQueryOptimizer.get_document_annotations` analysis check, `RelationshipQueryOptimizer.get_document_relationships` analysis check, `MetadataQueryOptimizer.get_documents_metadata_batch` corpus check) now pass `request=context` so the Tier-2 request-scoped cache (PR #1665) fires for the optimizer's existing analysis/corpus dereferences. The 21 other migrated sites live in helpers without a GraphQL request in scope (`check_analysis_permission`, `get_visible_extracts`, `_compute_effective_permissions` on the extract path, etc.) and degrade gracefully to Tier-1 instance caching. The annotation optimizer's bespoke `context._effective_perms_cache` (Phase A) is untouched — it continues to sit on top of `permission_cache` / `permission_optimizer`.
   - **Behavioral preservation.** The MIN(doc, corpus) bodies in both `_compute_effective_permissions` methods are unchanged; only the per-call shim invocations move. BACON-mode (`corpus.allow_comments`) branch ordering in the annotation optimizer is preserved verbatim per the issue's pin. The legacy `include_group_permissions=True` argument drops because `obj.user_can(...)` defaults to `True` (matching every other `Manager.user_can` / `obj.user_can` surface, and matching what these call sites were already passing explicitly).
