+++ config/graphql/document_mutations.py
@@ -1141,7 +1141,7 @@ class RestoreDeletedDocument(graphene.Mutation):
     """
     Restore a soft-deleted document path within a corpus.
 
-    Delegates to DocumentFolderService.restore_document() for:
+    Delegates to CorpusObjsService.restore_document() for:
     - Permission checking (corpus UPDATE permission)
     - Creating new DocumentPath with is_deleted=False
     """
@@ -1161,7 +1161,9 @@ class RestoreDeletedDocument(graphene.Mutation):
     @login_required
     @graphql_ratelimit(rate=RateLimits.WRITE_MEDIUM)
     def mutate(root, info, document_id, corpus_id) -> "RestoreDeletedDocument":
-        from opencontractserver.corpuses.folder_service import DocumentFolderService
+        from opencontractserver.corpuses.corpus_objs_service import (
+            CorpusObjsService,
+        )
 
         user = info.context.user
         not_found_msg = "Document or corpus not found, or you do not have permission."
@@ -1202,7 +1204,7 @@ class RestoreDeletedDocument(graphene.Mutation):
                 )
 
             # Delegate to service - handles permission checks and restoration
-            success, error = DocumentFolderService.restore_document(
+            success, error = CorpusObjsService.restore_document(
                 user=user,
                 document_path=deleted_path,
                 request=info.context,
@@ -1258,7 +1260,9 @@ class PermanentlyDeleteDocument(graphene.Mutation):
     @login_required
     @graphql_ratelimit(rate=RateLimits.WRITE_MEDIUM)
     def mutate(root, info, document_id, corpus_id) -> "PermanentlyDeleteDocument":
-        from opencontractserver.corpuses.folder_service import DocumentFolderService
+        from opencontractserver.corpuses.corpus_objs_service import (
+            CorpusObjsService,
+        )
 
         user = info.context.user
         not_found_msg = "Document or corpus not found, or you do not have permission."
@@ -1278,7 +1282,7 @@ class PermanentlyDeleteDocument(graphene.Mutation):
             except Corpus.DoesNotExist:
                 return PermanentlyDeleteDocument(ok=False, message=not_found_msg)
 
-            success, error = DocumentFolderService.permanently_delete_document(
+            success, error = CorpusObjsService.permanently_delete_document(
                 user=user,
                 document=document,
                 corpus=corpus,
@@ -1320,7 +1324,9 @@ class EmptyTrash(graphene.Mutation):
     @login_required
     @graphql_ratelimit(rate=RateLimits.WRITE_MEDIUM)
     def mutate(root, info, corpus_id) -> "EmptyTrash":
-        from opencontractserver.corpuses.folder_service import DocumentFolderService
+        from opencontractserver.corpuses.corpus_objs_service import (
+            CorpusObjsService,
+        )
 
         user = info.context.user
 
@@ -1330,7 +1336,7 @@ class EmptyTrash(graphene.Mutation):
             # caller; service layer enforces write/DELETE permission afterwards
             corpus = Corpus.objects.visible_to_user(user).get(pk=corpus_pk)
 
-            deleted_count, error = DocumentFolderService.empty_trash(
+            deleted_count, error = CorpusObjsService.empty_trash(
                 user=user,
                 corpus=corpus,
                 request=info.context,
