+++ config/graphql/document_types.py
@@ -1141,16 +1141,18 @@ class DocumentType(AnnotatePermissionsForReadMixin, DjangoObjectType):
         """
         Get folder assignment for this document in a specific corpus.
 
-        Delegates to DocumentFolderService.get_document_folder() for
+        Delegates to CorpusObjsService.get_document_folder() for
         permission checking and dual-system consistency.
         """
-        from opencontractserver.corpuses.folder_service import DocumentFolderService
+        from opencontractserver.corpuses.corpus_objs_service import (
+            CorpusObjsService,
+        )
         from opencontractserver.corpuses.models import Corpus
 
         _, corpus_pk = from_global_id(corpus_id)
         try:
             corpus = Corpus.objects.get(pk=corpus_pk)
-            return DocumentFolderService.get_document_folder(
+            return CorpusObjsService.get_document_folder(
                 user=info.context.user,
                 document=self,
                 corpus=corpus,
